AudioLinkVis: parametrise functions on how many samples to use
We can now select how many samples of PCM data to use for each of the
function. Particularly neccessary for the get_value_xy and
get_value_xy_scatter, as they are quite slow if given the full 2048
samples.

I also find that get_value_circle looks subjectively better at a small
amount of samples like 128 or so.