AudioLinkVis: Add some chronotensity-based scrolling and scaling
This includes some options that may or may not be used specifically to
work with ShaderFes 2021 sliders (I'm probably not going to use all of
them), and might be removed later because they are redundant.

Another change was made to apply the tiling parameters so that we
scale around the center of UV. This currently applies across the
board, which might maybe be a bit surprising for the _ST parameter.