XZALVis: Add random rotation reversing (experimental)
It's not quite as good as I want it to be. it still snaps a bit too
much...

Needs tweaking. It is currently a bit too jumpy.

Note that this also works without enabling chronotensity rotation.
It'll affect it even when we have a fixed rotation, but no
chronotensity rotation.