XZALVis: Move audiolink helper funcs into a header
With this we also move the application of _Amplitude_Scale to outside
of these helper functions, so these functions can be reused
independently of that global variable. We also stop applying
_Amplitude_Scale to DFT data, as that is scaled with the gain value
set in audiolink anyway.

TODO: Include the upstream audiolink header in this project, and
      remove the parts in my own header that duplicates that.

TODO: maybe try having auto-gain for PCM and we won't need _Amplitude_Scale?