Make a projector version of XZAudioLinkVisualizer
A quick hack based on the excellent projector gimmick available on
Booth (which thankfully is MIT licensed so I can use it), but with my
shader quickly hacked in.

The shaft portion is too laggy currently so it is disabled by default.
I might either forego it entirely or perhaps use suitably
low-res CRT (512x512 ~ 1024x1024) for the shaft portion, but for the
projector use the current infinite-res version.