XZAlVis: Some tweaks to get_value_xy_* function
Most importantly: Use 1.#INF to get infinity. The 1.0/0.0 thing
produces warnings when compiling.

Also we remove the ability to specify the amount of samples as a
parameter to the function in the hopes that it might optimize a smidge
better (for now though I do not go as far as explicitly commanding an
    unroll of the loop).

TODO: Unsure if the compiler gets limited by the fact that the loop
appears at first to be always dependent on the previous step. Perhaps
we should give hints by manually unrolling the loop a little. Say we
compute four min values inside the loop proper and then find the
minimum of that after the loop.