XZAlVis: Attempt to optimize the XY plots by manual unrolling
My guess is that the compiler will have trouble even if told to unroll
these loops because of the data dependency on the previous loop when
using the min instruction. We use float4s here to unroll the loop 4
times, and coming out of the loop we will have 4 minimum values that
we can get the final min value from.