WIP: Another XZAlVis CRT experiment started
This time trying to make one of those ever-present scaling and
rotating trails. Possible since we are now writing the shader
specifically for CRTs.

TODO: * make the trail type and such configurable
      * chronotensity-driven trail
      * get rid of the silly updateszone things. That's just me
        playing around right now.
      * Add symmetry/kaleidoscope features to XZAudioLinkVisualizer.cginc itself

Further fields of study: Investigate whether we could make the trail
thing a separate shader that can add do trails from another input
shader set as the initializer material in a CRT. This would be more modular.