XZAlVis: Optimize IR output size
Turns out that the inliner in the compiler is a lot dumber than I
thought, and the multiple paths I take that eventually call get_color
end up with the code for that portion (easily recognizeable by the
switch...case) duplicated at least 3 times.

Fix this by rewriting handling of Auto and Auto2 modes in a dumber
way... Actually maybe this way is better anyway. It is less lines of
code, at least...