XZAudioLinkGeometrVectorScope: Now mimics the lok of XZAlVis
Produce quads and draw a dot on it using the line drawing function we
have used before (slightly tweaked so we do not need to have too big
    quads)

We also made it move a bit in 3D by applying mono PCM data there.
TODO: make a toggle for that and also make the quads billboard when it
is active.

TODO: Find a shorter name..
      Perhaps use a texture instead for optimization (it is pretty
    slow when you are up close)