XZAlVis: Add a global parameter for UV tiling and offset _ST0
This is because sometimes we just want to make the shader "fit" on an
object with e.g. skewed UV and such.

Meanwhile the old _ST is often used to say make the shader show up
five times on a quad that has been stretched to be rectangular, all
the while wanting to have the vignette just fade out towards the edges
of that quad.

For now this is only available in the basic XZAlVis version, as there
isn't much point in expanding it to the other ones for now.