XZAlVis: Add a quick hack that can do some neat tricks based on vertex color
This is to be used on the hexagon wall in the vket2023 summer booth.
It is quite the hackity hack... in particular the mode mixing upping
is not pretty.

We also include a material set up for this particular task.
We do not include any model to use it with yet though, but it might at
least serve as some sort of example?