RaymarchedSomething3D: Some random messing about
The code is messy indeed. I will have to clean it up at some point.
I think using out parameters with structs or something might make for
a cleaner look or whatever. It could include things like the rotation
matrix and the translation vector for the object. And UV etc. etc.