(WIP) Start working on a controller for the XZAudioLinkVisualizer series
Start working on this again. Eventually I would like some more
automated approach that can generate prefabs on the fly or so that
let's you control material properties in a nice synchronized manner.

Compared with my old solution I do not need a separate udon behavior
just to store the synced variables and such, that was a bit wasteful
and the result of me just copying audiolink before I understood Udon
better.

As of this commit this isn't completely working yet, the preview
planes don't work, but some of the settings do seem to apply?