XZAlVisController: More tweaks
In particular we hard-code the values for the blend modes for the
active and preview view on the controller in case the material we
control has some other settings fit for its peculiar setup that might
just result in them being completely black on the controller. (e.g. XZAlVis Shaft CRT in.mat)