AudioLinkRing: Significantly tweak how it works
We now have a separate cooldown time during which the ring can be
retriggered, and the other time value is just a measure of how
"fast" the ring should move outwards.

There is also an upper and a lower threshold value now, with
retriggering requiring a higher threshold in the normal case, and in
the hold case we require going below the lower threshold before
releasing the ring.