RenderRing: Add some alternative rings
Some nice spirally ones and stuff. I should probably name them
something more easy to remember and stuff, and weed out a few of them
eventually. But for now it's like this.

We also introduce another value in RingCRT here, where we use .a just
the same for counting upwards but it actually remembers its value
while the ring is travelling outwards.