RenderParticles: Remove the need for the _NumPoints input parameter
Because of course I just needed to be more creative with how I
calculate the operationID. Argh...

Now you only need to tweak the input mesh and the number of instances
to get your desired max amount of particles.

The amount of loops could perhaps be made configurable too in the
future, but doing it as an input parameter would make the loop not as
easily unrollable. Doing it as keywords, like with the instances,
would certainly be possible, but we already have quite a bit of
keywords.