ParticleCRTGun WIP
Currently fails with "cannot use casts on l-values" on the
emit_particles call.
I'm not sure why that is because I don't think I'm doing any casting
here?

This is just a commit so I can change things in the code without fear.

ParticleCRTGun: Emit 16 particles on bass

ParticleCRTGun: Fix in the angle calculation

ParticleCRTGun.mat: Some wierd change unity did automagically

ParticleCRTGun: Whitespace fix

Renaming of the ParticleCRTGun material in

try emit_particle without the emit_context

This seems to work, but it is definitely a bit ugly I think. Ugh...

Half-way house of only parameters being modified in the struct seems ok

Why?

Turns out it works if you do it this way for some reason

It seems like that main problem was not initializing the whole struct.
Huh... In C it would be UB of course, but I thought HLSL was smart
enough to just consider structs always 0-initialized.