Create ParticleCRTGun, an audiolink gun particle effect of some sort
Sadly I can't figure out a way to have the particles move
independently of the gun, at least in an avatar. It would require a
way to find out the position of the gun from the CRT shader. If it
wasn't a CRT shader we could probably (ab)use light sources or GPU
instancing.

However it still looks pretty cool as is. Maybe I'll try
making a world variant where you waving the gun around only affects
already shot particles by moving the axis to which they get attracted
to and so.

TODO:  * Make the bounds larger so it works when not looking straight on
         as well (either create a variant of Points64 with larger bounds, or
         use a SkinnedMeshRenderer)
       * Play around and tweak with it more/make other variants?
       * Refactor the emit_particle function out of ParticleCRTGun as
         it is pretty general already (make _AlwaysEmit a compile-time
         define perhaps?)

Notes: The the starting point when we are forced to overwrite
       particles is randomized. This should likely be more visually
       pleasing than starting always from idx 0.