[WIP] Start creating RaymarchedPalmTree shader
This is far from being a palmtree right now... Although I do want to
attempt that eventually...

For now it's just me experimenting with refactoring stuff from
RaymarchedTunnel by using more advanced HLSL stuff like classes and
interfaces...

Material tweak

Attempts to refactor the getting of color

It doesn't work so well, we end up always calling the GetColor
placeholder in the base SDFObject class making them yellow.

Compilation failed: interface references must resolve to non-varying objects

Seems my current approach is not working out so well.
Maybe I should go the full closures-based approach?

Or maybe I should give up and use some mixed approach using macros to
some extent.

Or give up even more and just use structs and switch .. case or so...

Simplify down to just one interface again

At least this sort of kind of words, but the implementation of
GetColor in SceneSDF is unfortunately not so nice and clean...

TODO: Make a chainable thingy that does rotation and translation on a
given SDF so it isn't hard-coded into them like now?

Change GetColor signature

It just takes the position we intersected at and the marching
direction. It is up to the SDF object itself whether it needs to
compute the normal or UV or things like that.

misc changes

brighter red cube, larger moon

Making something more composable using the closure hack

Unfortunately we cannot "recurse" (even though it would be possible
    to find a static solution, but HLSL is not smart enough), so we
require hard-coded variants of the MinSDFx class. Currently
it is implemented for up to 4 SDFs, planning to go up to 8 I guess.

Fix TranslateSDF

Find problem: Somehow values I set in the static constructors become global

Etc. the tint color setting in BoxSDF or having more than one RotateSDF

The currently set one we should see two cubes and one sphere, but the
cube that is supposed to be in the middle just rotating is nowhere to
be found.

Now it sorta works, but probably only because I have two different rotaters

Yes indeed like the previous commit said

Even when I did the straight-up normal function to return a closure it
only works if I duplicate the function...

How annoying... That means that this method of composition will not be
suitable for handling rotation and translation... I guess it could
work if we build around always having a transform, like I was
attempting a few commits back.

Considering that, perhaps only the
interfaces-as-function-pointers-trick is all we need too. Mostly to
make a reusable shortestDistanceToSurfaceWithColor function, reusable
EstimateNormal etc.

Speaking of normals, we also estimate them locally to the object now,
that is a bit of an issue too.

Apparently forgoing using the static method contsructors works properly

Lolwhy...

Although this seems semi-usable like this now. I think the syntax
could be prettified with some simple macros to make this approach more
viable, although unfortunately not quite as clean of course.

Changing the return type to be more explicit in the constructor seems to help?

For reference this did not work properly when BoxSDF::New returned
ISDFObject (and box1, box2, box3 where also all declared as such)

Seems like only the base objects can have constructors that return their own type

So BoxSDF::New can return BoxSDF, but if I try it with
TranslateSDF::New it will result in Next() not being overriden (it
    will be the base version in SDFObjectBase which right now makes it
    a magenta sphere)

Basically we will always have to go the route of always declaring new
classes locally (with some macros thrown in to make it more palatable,
    probably).

Cleanup some and play around with UV mapping

The noise texture is now a wood texture (outside of the repo so it
    only works in the current project I'm on)

Switching to using magical macros for composition

This way we always create a truly local class, which seems to lead to
better-working closures. (Seeing how it worked before, HLSL seems to
    have some version of the upwards-funargs problem?)

Cleanup

Maybe better reflections?

No idea...

Now the still remaining problem though is that the normal gets
computed object-locally. We should probably compute it in
shortestDistanceToSurfaceWithColor and pass it into GetColor or
something like that.

Alternatively just make all objects have a transform and be
self-responsible or so (See: how SphereSDF functions currently)

Move normal calculation to top level

This fixes reflections and the shadows (TODO: make better proper
    shadows with support for lights and stuff)

Downside is that it seems a bit superfluous in the case that the
normal is not needed (even though in theory the compiler could figure
    that out given everything gets inlined, but I doubt it will given
    its past performance...)

Calculating the normal in itself is potentially a bit more expensive
now too, since it needs to regard all the objects when done at the top
level instead of just the one we know we hit.

Further doing the unrotation and translation etc. of the point is now
also superfluous unless you are getting UV for the object. The
compiler could figure out that computing that is unneccessary though,
given all the inlining, but it is sometimes a lot more stupid than
you'd expect.

messing about

SimplexNoiseTexture: Make a bit more dense or so

RaymarchedPalmTree: Get rid of needing to put _SceneScale everywhere

This should of course be done globally.
However, note how I also had to adjust EPSILON and MAX_DIST etc.
Perhaps it'd be best if these values also scaled along with
_SceneScale? Woudl require some slight rearchitecting though...
And it's going to be even harder once I break out a lot of these
functions into a header.