Yes indeed like the previous commit said
Even when I did the straight-up normal function to return a closure it
only works if I duplicate the function...

How annoying... That means that this method of composition will not be
suitable for handling rotation and translation... I guess it could
work if we build around always having a transform, like I was
attempting a few commits back.

Considering that, perhaps only the
interfaces-as-function-pointers-trick is all we need too. Mostly to
make a reusable shortestDistanceToSurfaceWithColor function, reusable
EstimateNormal etc.

Speaking of normals, we also estimate them locally to the object now,
that is a bit of an issue too.