Move normal calculation to top level
This fixes reflections and the shadows (TODO: make better proper
    shadows with support for lights and stuff)

Downside is that it seems a bit superfluous in the case that the
normal is not needed (even though in theory the compiler could figure
    that out given everything gets inlined, but I doubt it will given
    its past performance...)

Calculating the normal in itself is potentially a bit more expensive
now too, since it needs to regard all the objects when done at the top
level instead of just the one we know we hit.

Further doing the unrotation and translation etc. of the point is now
also superfluous unless you are getting UV for the object. The
compiler could figure out that computing that is unneccessary though,
given all the inlining, but it is sometimes a lot more stupid than
you'd expect.