RaymarchedPalmTree: Get rid of needing to put _SceneScale everywhere
This should of course be done globally.
However, note how I also had to adjust EPSILON and MAX_DIST etc.
Perhaps it'd be best if these values also scaled along with
_SceneScale? Woudl require some slight rearchitecting though...
And it's going to be even harder once I break out a lot of these
functions into a header.