ParticleCRTGun: Break out the particle emit helper functions
These are quite nicely reusable. As a compromise though, we have to do
the toggling of whether to always emit (by overwriting) or not a
define. An upside now though is that if you turn that off it should
result in somewhat simpler code and potential speed gains (probably
all but negligible)