WorldSpaceGun now works with rotation
We change to simply drawing UNITY_MATRIX_M out to the grabpass, and
then using that in ParticleCRTGunWorldSpace to do translation and
rotation. Scaling is cancelled out by normalizing the row-vectors, and
the translation is removed by converting from float4x4 to float3x3
effectively removing the last row of the matrix.