ParticleCRTGunWorldSpace: Increase bounding box of the GrabPass
To have less isues when it doesn't become visible. Since the shader
relies on the input being a quad, let's just use a SkinnedMeshRenderer
for now. In the future we could make a quad mesh with different bounds
however, just like we created the particle point meshes with larger
bounds.