WIP work on a CRT that receives pixels on a clock signal thing
Currently just a PoC and it currently writes lines as it writes every
frame while _WR is high. We should probably change it instead to
trigger every time _WR changes (DDR, both flanks).

Currently I am just testing it using a C# script, but eventually we
are going to have to make it accept input from a GrabPass instead if
we want it to be usable in avatars, as we cannot control the CRT
material with animators. The purpose of the GrabPass shader will be
simply to pass parameters on from OSC input on the avatar to the CRT
update material.