PixelSendCRT: Refactor magic value 16 into BYTES_PER_SEND
This is going to make it easier later when we might want to extend the
amount of bytes sent at once.

Although for the direct-parameter version we are still hard-coded, so
add a _Static_assert for now. We could make it variable (up to a
    hard-coded maximum) in the future with some macro magic to exclude
parameters above BYTES_PER_SEND